Server_Side/requirements.txt
asgiref==3.3.1
astroid==2.4.2
boto3==1.16.47
botocore==1.19.47
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-extensions==3.1.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.0.1
pydotplus==2.0.2
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.50.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1